Remove old hacks and handle end of audio correctly.
The correct way in Qt to find out when audio playback has finished is to
connect to the stateChanged signal of the QAudioOutput and look for a
transition to QAudio::IdleState.

It seems that in the past, this didn't always work correctly on some
platforms. As a workaround, CuteCW was using its own 'generatorDone'
signal as an alternative way to get an event at the end of playback.

However, the generatorDone signal only indicates when the last sample
has been read from the generator, not when it has actually been played
out by the audio device. The audio driver may read samples from the
input to buffer them well before they are actually played. As such,
users have sometimes encountered playback being truncated. There have
been some attempts to fix this by adding padding to the buffer, but this
is not a reliable approach.

This commit simply assumes that the old platform specific bugs are long
gone by now. It uses the correct way to handle end of playback, and gets
rid of the generatorDone signal and other related tricks.

This change fixes truncated playback for me with Qt 5.11 on Linux. It
should work for all other platforms too, as long as Qt works correctly.